requirements.txt
alembic==0.9.8
amqp==2.2.2
appnope==0.1.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Assets==0.12
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.3.3
Flask-CDN==1.5.3
Flask-Celery==2.4.3
Flask-Celery-Helper==1.1.0
Flask-Cors==3.0.3
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
freetype-py==1.2.1
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
parso==0.1.1
pep8==1.7.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyMySQL==0.8.0
pytesseract==0.2.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2018.3
qiniu==7.2.0
redis==2.10.6
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.4
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
visitor==0.1.3
wcwidth==0.1.7
webassets==0.12.1
Werkzeug==0.14.1
WTForms==2.1
pytesseract==0.2.0
moviepy==0.2.3.2
opencv-python==3.4.0.12
scipy==1.0.1
qrcode==6.0
